feat: migrate `@boundary` annotation (#1146)
Closes #1142.

### Summary of Changes

- add migration for `@boundary` annotations to another version, with the
possibilities:
- If the mapped parameter has a number boundary and its type is known,
the annotation will be migrated as expected. The interval would be
converted if the type of the parameter from apiv1 and the type of the
parameter from apiv2 mismatch (int, float). From a floating point
boundary to an integer boundary, the boundary would be smaller. E.g.,
The interval (0.5, 9.5) will be converted to [1, 9]. If a conversion
happens, the migrated annotation will be marked as unsure.
- If the mapped parameter is a parameter, but its type in unknown, an
unsure boundary annotation will be created for this parameter, but
without any conversion.
- If the element is not a parameter, a `@todo` annotation will be
created.


### Testing Instructions

run the migrate command or view and run the test_migration.py file

Co-authored-by: Aclrian <Aclrian@users.noreply.github.com>