feat: migrate `@value` annotation (#1148)
Closes #1145.

### Summary of Changes

- add migration for `@value` annotations to another version, with the
possibilities:
- If the variant of the annotation is omitted and both parameter have
the same value, the annotation will be migrated as expected. If the type
is different, the annotation will be marked as unsure.
- If the variant of the annotation is required and the parameter from
apiv2 has a default value of the same type, the annotation will be
migrated as expected.
- If the variant is optional or constant and the type of the parameter
from apiv2 is known and the type of the parameters match, an annotation
will be migrated as expected. If the type is unknown, the annotation
will be marked as unsure.

### Testing Instructions

run the migrate command or view and run the `test_migration.py` file

Co-authored-by: Aclrian <Aclrian@users.noreply.github.com>