ci: remove merge queue workflows (#1280)
### Summary of Changes

We have some checks that should be a requirement before PRs can be added
to a merge queue:

* dependency review
* check title
* lint

Naturally, the code should also build. In the merge queue itself, we
only need to verify again that the code still builds as previous PRs get
merged in. Checking the title of the PR again, for example, is a waste
of time since it cannot have changed.

However, this cannot be realized since checks that are marked as
required must be run in the merge queue and on pull request events.

Conclusion: With our current throughput, a merge queue just slows us
down.